# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: AI IT Helpdesk Agent
  description: AI IT Helpdesk Agent backend with chat, ticketing, and HR request APIs.
  version: 0.1.0
paths:
...
  /api/v1/tickets:
    post:
      tags:
      - Tickets
      summary: Create an IT support ticket
      description: Create a new helpdesk ticket.
      operationId: create_api_v1_tickets_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TicketCreateRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - Tickets
      summary: List tickets for an employee
      description: List tickets optionally filtered by employee ID.
      operationId: read_many_api_v1_tickets_get
      parameters:
      - name: employee_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Employee Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/tickets/{ticket_id}:
    get:
      tags:
      - Tickets
      summary: Fetch a ticket by ID
      description: Retrieve a single ticket by identifier.
      operationId: read_api_v1_tickets__ticket_id__get
      parameters:
      - name: ticket_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Ticket Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/tickets/{ticket_id}/status:
    patch:
      tags:
      - Tickets
      summary: Update a ticket status
      description: Update the ticket lifecycle status.
      operationId: update_status_api_v1_tickets__ticket_id__status_patch
      parameters:
      - name: ticket_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Ticket Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TicketStatusUpdateRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/tickets/{ticket_id}/escalate:
    post:
      tags:
      - Tickets
      summary: Escalate a ticket to a human agent
      description: Escalate a ticket for human review.
      operationId: escalate_api_v1_tickets__ticket_id__escalate_post
      parameters:
      - name: ticket_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Ticket Id
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/software/request:
    post:
      tags:
      - Software Requests
      summary: Submit a software request
      description: Create a software procurement or access request.
      operationId: request_api_v1_software_request_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SoftwareRequestCreateRequest'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SoftwareRequestResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/software/request/{request_id}/status:
    get:
      tags:
      - Software Requests
      summary: Get software request status
      description: Retrieve the status for a software request.
      operationId: request_status_api_v1_software_request__request_id__status_get
      parameters:
      - name: request_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Request Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SoftwareRequestStatusResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    MessageResponse:
      properties:
        message:
          type: string
          title: Message
      additionalProperties: false
      type: object
      required:
      - message
      title: MessageResponse
...
    SoftwareRequestCreateRequest:
      properties:
        employee_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Employee Id
        requester_email:
          anyOf:
          - type: string
            format: email
          - type: 'null'
          title: Requester Email
        software_name:
          type: string
          maxLength: 200
          minLength: 1
          title: Software Name
        justification:
          type: string
          maxLength: 2000
          minLength: 1
          title: Justification
        business_impact:
          type: string
          maxLength: 100
          title: Business Impact
          default: standard
      additionalProperties: false
      type: object
      required:
      - software_name
      - justification
      title: SoftwareRequestCreateRequest
    SoftwareRequestResponse:
      properties:
        request_id:
          type: string
          format: uuid
          title: Request Id
        status:
          type: string
          enum:
          - requested
          - under_review
          - approved
          - rejected
          - fulfilled
          title: Status
        software_name:
          type: string
          title: Software Name
        employee_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Employee Id
        requester_email:
          anyOf:
          - type: string
            format: email
          - type: 'null'
          title: Requester Email
        created_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Created At
        note:
          anyOf:
          - type: string
          - type: 'null'
          title: Note
      additionalProperties: false
      type: object
      required:
      - request_id
      - status
      - software_name
      title: SoftwareRequestResponse
    SoftwareRequestStatusResponse:
      properties:
        request_id:
          type: string
          format: uuid
          title: Request Id
        status:
          type: string
          enum:
          - requested
          - under_review
          - approved
          - rejected
          - fulfilled
          title: Status
        note:
          anyOf:
          - type: string
          - type: 'null'
          title: Note
      additionalProperties: false
      type: object
      required:
      - request_id
      - status
      title: SoftwareRequestStatusResponse
...
    TicketCreateRequest:
      properties:
        employee_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Employee Id
        requester_email:
          anyOf:
          - type: string
            format: email
          - type: 'null'
          title: Requester Email
        subject:
          type: string
          maxLength: 200
          minLength: 1
          title: Subject
        description:
          type: string
          maxLength: 10000
          minLength: 1
          title: Description
        category:
          type: string
          maxLength: 100
          title: Category
          default: general
        priority:
          type: string
          enum:
          - low
          - medium
          - high
          - urgent
          title: Priority
          default: medium
      additionalProperties: false
      type: object
      required:
      - subject
      - description
      title: TicketCreateRequest
    TicketListResponse:
      properties:
        tickets:
          items:
            $ref: '#/components/schemas/TicketResponse'
          type: array
          title: Tickets
      additionalProperties: false
      type: object
      title: TicketListResponse
    TicketResponse:
      properties:
        ticket_id:
          type: string
          format: uuid
          title: Ticket Id
        employee_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Employee Id
        status:
          type: string
          enum:
          - new
          - open
          - in_progress
          - resolved
          - closed
          title: Status
        subject:
          type: string
          title: Subject
        requester_email:
          anyOf:
          - type: string
            format: email
          - type: 'null'
          title: Requester Email
        category:
          anyOf:
          - type: string
          - type: 'null'
          title: Category
        priority:
          anyOf:
          - type: string
          - type: 'null'
          title: Priority
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        created_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Created At
      additionalProperties: false
      type: object
      required:
      - ticket_id
      - status
      - subject
      title: TicketResponse
    TicketStatusUpdateRequest:
      properties:
        status:
          type: string
          enum:
          - new
          - open
          - in_progress
          - resolved
          - closed
          title: Status
      additionalProperties: false
      type: object
      required:
      - status
      title: TicketStatusUpdateRequest
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: /api/v1/auth/login
...